chore: untangle the react related exports
