+++ packages/compass-preferences-model/src/read-only-preferences-access.ts
@@ -1,0 +1,48 @@
+import { createNoopLoggerAndTelemetry } from '@mongodb-js/compass-logging/provider';
+import { Preferences, type PreferencesAccess } from './preferences';
+import { type AllPreferences } from './preferences-schema';
+import { InMemoryStorage } from './in-memory-storage';
+
+export class ReadOnlyPreferenceAccess implements PreferencesAccess {
+  private _preferences: Preferences;
+  constructor(preferencesOverrides?: Partial<AllPreferences>) {
+    this._preferences = new Preferences({
+      logger: createNoopLoggerAndTelemetry(),
+      preferencesStorage: new InMemoryStorage(preferencesOverrides),
+    });
+  }
+
+  savePreferences() {
+    return Promise.resolve(this._preferences.getPreferences());
+  }
+
+  refreshPreferences() {
+    return Promise.resolve(this._preferences.getPreferences());
+  }
+
+  getPreferences() {
+    return this._preferences.getPreferences();
+  }
+
+  ensureDefaultConfigurableUserPreferences() {
+    return this._preferences.ensureDefaultConfigurableUserPreferences();
+  }
+
+  getConfigurableUserPreferences() {
+    return Promise.resolve(this._preferences.getConfigurableUserPreferences());
+  }
+
+  getPreferenceStates() {
+    return Promise.resolve(this._preferences.getPreferenceStates());
+  }
+
+  onPreferenceValueChanged() {
+    return () => {
+      // noop
+    };
+  }
+
+  createSandbox() {
+    return Promise.resolve(new ReadOnlyPreferenceAccess(this.getPreferences()));
+  }
+}
